+++ online-food-ordering/src/main/java/com/example/online/food/ordering/model/OrderedItems.java
@@ -16,6 +16,6 @@ public class OrderedItems {
     @ManyToOne
     private Foods foods;
     private int quantity;
-    private Long totalPrice;
+    private int totalPrice;
     private List<String> ingredients;
 }
